+++ NEWS.md
@@ -1,5 +1,14 @@
 # MsBackendMetabolights 0.0
 
+## Changes in 0.0.3
+
+- Add vignette and `backendMerge,MsBackendMetaboLights` function.
+
+## Changes in 0.0.2
+
+- Add `MsBackendMetaboLights` class, constructor and `backendInitalize()`
+  method.
+
 ## Changes in 0.0.1
 
 - Add utility functions to retrieve information from MetaboLights.